Windows: Fix type conversion warning with 64-bit MSVC.
request_handler_unittest.cc(448,55): warning C4267: 'argument':
conversion from 'size_t' to 'int', possible loss of data

+++ tests/ceftests/request_handler_unittest.cc
@@ -451,7 +451,7 @@ class NetNotifyRendererTest : public ClientAppRenderer::Delegate,
 void RunNetNotifyTest(NetNotifyTestType test_type,
                       bool same_origin,
                       size_t count = 3U) {
-  TestHandler::CompletionState completion_state(count);
+  TestHandler::CompletionState completion_state(static_cast<int>(count));
   TestHandler::Collection collection(&completion_state);
 
   std::vector<CefRefPtr<NetNotifyTestHandler>> handlers;
